deleted refactored feature processing classes, optimized corresponding imports
