+++ ptls/data_load/data_module/cpc_data_module.py
@@ -20,7 +20,7 @@ from ptls.data_load import IterableAugmentations, IterableChain, padded_collate_
 from ptls.data_load.augmentations.build_augmentations import build_augmentations
 from ptls.data_load.data_module.map_augmentation_dataset import MapAugmentationDataset
 from ptls.data_load.datasets.parquet_dataset import ParquetDataset, ParquetFiles
-from ptls.data_load.iterable_processing.feature_filter import FeatureFilter
+from ptls.data_load.iterable_processing import FeatureFilter
 from ptls.data_load.iterable_processing.filter_collection import CategorySizeClip, SeqLenFilter
 from ptls.data_load.iterable_processing.iterable_shuffle import IterableShuffle
 from ptls.data_load.list_splitter import ListSplitter
