+++ ptls/data_load/iterable_processing/__init__.py
@@ -1,9 +1,6 @@
-from .feature_rename import FeatureRename
-from .feature_bin_scaler import FeatureBinScaler
-from .feature_filter import FeatureFilter
-from .feature_type_cast import FeatureTypeCast
 from .filter_collection import (IdFilter, SeqLenFilter, CategorySizeClip, ISeqLenLimit,
                                 FilterNonArray, FilterNonArray, IdFilterDf)
+from .processing_collection import FeatureBinScaler, FeatureFilter, FeatureRename, FeatureTypeCast
 from .iterable_shuffle import IterableShuffle
 from .take_first_trx import TakeFirstTrx
 from .target_empty_filter import TargetEmptyFilter
