+++ ptls_tests/test_data_load/test_iterable_processing/test_seq_len_filter.py
@@ -1,4 +1,4 @@
-from ptls.data_load.iterable_processing.filter_collection import SeqLenFilter
+from ptls.data_load.iterable_processing import SeqLenFilter
 import numpy as np
